Migrate ChainServlet to package org.apache.commons.chain.web, since it will be useful in portlet and JavaServer Faces environments as well.
Add ChainListener as an alternative configuration mechanism for
Servlet 2.3 or later systems (it takes advantage of the getResourcePaths()
method on ServletContext to scan all JAR files in "/WEB-INF/lib" for
"META-INF/chain-config.xml" resorces, and loads them automatically).

Tweak the build.xml conditionals to correctly represent the fact that
Digester depends on BeanUtils and Collections also being present.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/chain/trunk@142819 13f79535-47bb-0310-9956-ffa450edef68