Make CatalogFactory an abstract class instead of an interface, and make the getInstance() and clear() methods static on this class.  This is necessary both to avoid applications having to reference CatalogFactoryBase, but also to ensure that the clear() method actually releases all references to the relevant class loader -- in the original scheme, we would still have been leaving garbage lying around.
Refine ConfigRuleSet so that multiple <factory> elements with the same
catalog name will cause the nested chain and command definitions to be
merged, instead of replacing the entire catalog with only the contents
of the last <factory> element.

FIXME: Modify ConfigureListener and ConfigureServlet to use the new
mechanisms that expect a <factory> element -- but find a backwards
compatible way to continue supporting existing use cases.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/chain/trunk@142900 13f79535-47bb-0310-9956-ffa450edef68