Make ContextBase live up to the Serializable contract that it inherits by virtue of extending HashMap, and add a unit test to prove it works. Fixed in nightly build 20050417.
However, this unit test pointed out a problem with the design of the
concrete WebContext subclasses (ServletWebContext, PortletWebContext,
FacesWebContext).  It is not going to be possible to make these classes
Serializable because the underlying container objects they wrap are not
Serializable -- and there is no way to un-inherit the "implements Serializable"
characteristic of the ContextBase base class.

PR:  Bugzilla #34409
Submitted By:  Jeff Ramsdale <jramsdale AT solutionsiq.com>


git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/chain/trunk@161600 13f79535-47bb-0310-9956-ffa450edef68